+++ src/emu/chan.c
@@ -15,8 +15,8 @@ chan_init(struct chan *chan, enum chan_type type, const char *fmt, ...)
 	va_list ap;
 	va_start(ap, fmt);
 
-	int n = ARRAYLEN(chan->name);
-	int ret = vsnprintf(chan->name, n, fmt, ap);
+	int n = (int) ARRAYLEN(chan->name);
+	int ret = vsnprintf(chan->name, (size_t) n, fmt, ap);
 	if (ret >= n)
 		die("channel name too long");
 	va_end(ap);
